+++ projects/ngx-collection/src/lib/tests/collection.sync.spec.ts
@@ -830,12 +830,12 @@ describe('Collection Service (sync)', () => {
     const items = coll.$items();
     expect(items).toStrictEqual([]);
     setTimeout(() => {
-      done();
       expect(handlerWasCalled).toBeTruthy();
       handlerWasCalled = false;
       const items = coll.$items();
       expect(handlerWasCalled).toBeFalsy();
       expect(items).toStrictEqual([item1, item2, item3]);
+      done();
     }, 1);
   });
 });